+++ src/color/coeff_spectrum.h
